requirements.txt
requests==2.31.0
streamlit==1.32.0
pandas==2.2.1
sqlalchemy==2.0.27
python-dotenv==1.0.1 